% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvency2rfr-package.R
\docType{package}
\name{solvency2rfr-package}
\alias{solvency2rfr}
\alias{solvency2rfr-package}
\title{solvency2rfr: EIOPA Risk-Free Interest Rate Term Structures for Solvency II}
\description{
Downloads and parses the risk-free interest rate (RFR) term structures published monthly by the European Insurance and Occupational Pensions Authority (EIOPA) for Solvency II calculations. Provides a tidy tibble interface to the data, accessed via the official EIOPA RSS feed at \url{https://www.eiopa.europa.eu/feed/53/rss_en}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JanWein/solvency2rfr}
  \item Report bugs at \url{https://github.com/JanWein/solvency2rfr/issues}
}

}
\author{
\strong{Maintainer}: Jan-Hendrik Weinert \email{janhendrikweinert@gmail.com} (\href{https://orcid.org/0009-0005-2071-6502}{ORCID})

Authors:
\itemize{
  \item Jan-Hendrik Weinert \email{janhendrikweinert@gmail.com} (\href{https://orcid.org/0009-0005-2071-6502}{ORCID})
}

}
\keyword{internal}
